Remove 'Last revised' line from acknowledgements
It complicates the generation of acknowledgements and can lead to
changes in the git history where only that value is updated and that are
therefore mainly noise without real benefit.
Because the acknowledgements are now generated and updated regularly and
automatically as part of the release process, the value of that field is
in general questionable.

To keep the workflow simpler respectively to not complicate it more in
order to avoid updates of only this field, it's just removed.